separated thunk in the services module
