+++ src/store/trending-reducer.js
@@ -1,5 +1,3 @@
-import getTrendingFeed from '../api/getTrendingFeed';
-
 const SET_TRENDING_FEED = '/trending/SET_TRENDING_FEED';
 const SET_ERROR = '/trending/SET_ERROR';
 
@@ -30,18 +28,5 @@ export const setTrendingFeed = (trendingFeed) => ({
     trendingFeed,
 });
 export const setError = (error) => ({ type: SET_ERROR, error });
-
-export const requestTrendingFeed = () => async (dispatch) => {
-    try {
-        const data = await getTrendingFeed();
-        if (data.length === 0 || data === 'something went wrong, please try again') {
-            dispatch(setError('Empty array trending feed'));
-        } else {
-            dispatch(setTrendingFeed(data));
-        }
-    } catch (err) {
-        dispatch(setError(err.response.data.message));
-    }
-};
 
 export default trendingReducer;
