+++ src/store/users-reducer.js
@@ -1,6 +1,3 @@
-import userFeedData from '../mocks/user-feed.json';
-import getUserInfo from '../api/getUserInfo';
-
 const SET_USERS_FEED = '/users/SET_USERS_FEED';
 const SET_USERS_INFO = '/users/SET_USERS_INFO';
 const TOGGLE_IS_FETCHING = 'users/TOGGLE_IS_FETCHING';
@@ -48,33 +45,5 @@ export const toggleIsFetching = (isFetching) => ({
     isFetching,
 });
 export const setRequestError = (error) => ({ type: SET_REQUEST_ERROR, error });
-
-// In case the data came from the server correctly
-/* export const requestUsersFeed = (id) => async (dispatch) => {
-    dispatch(toggleIsFetching(true));
-    const data = await getUserFeed(id);
-    dispatch(toggleIsFetching(false));
-    dispatch(setUsersFeed(data));
-} */
-
-export const requestUsersFeed = () => (dispatch) => {
-    const data = userFeedData;
-    dispatch(setUsersFeed(data));
-};
-
-export const requestUsersInfo = (id) => async (dispatch) => {
-    try {
-        dispatch(toggleIsFetching(true));
-        const data = await getUserInfo(id);
-        dispatch(toggleIsFetching(false));
-        if (!Object.keys(data).length || data === 'something went wrong, please try again') {
-            dispatch(setRequestError('Empty object in userInfo'));
-        } else {
-            dispatch(setUsersInfo(data));
-        }
-    } catch (error) {
-        dispatch(setRequestError(error.response.data.message));
-    }
-};
 
 export default usersReducer;
